chore: transform non-src files to typescript
