--- .babelrc
@@ -1,22 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "loose": true,
-        "modules": false
-      }
-    ]
-  ],
-  "plugins": [
-    "no-side-effect-class-properties",
-    [
-      "@babel/plugin-proposal-class-properties",
-      {
-        "loose": true
-      }
-    ],
-    "transform-es3-property-literals",
-    "transform-es3-member-expression-literals"
-  ]
-}
