+++ .gitignore
@@ -208,4 +208,5 @@ temp/
 demo/_data/version.yml
 demo/release/
 CHANGELOG.md
+declaration/
 dist/
