+++ package.json
@@ -4,14 +4,14 @@
 	"description": "A base class utility that manages events and options in modules.",
 	"main": "dist/component.js",
 	"module": "dist/component.esm.js",
-	"types": "src/component.d.ts",
+	"types": "declaration/Component.d.ts",
 	"scripts": {
-		"start": "webpack-dev-server --open",
-		"build:webpack": "webpack --env production",
-		"build": "rollup -c",
+		"build": "rm -rf ./dist ./declaration && rollup -c && npm run declaration && npm run printsizes",
+		"declaration": "rm -rf declaration && tsc -p tsconfig.declaration.json",
+		"printsizes": "print-sizes ./dist --exclude=\\.map",
 		"test": "karma start",
 		"test:chrome": "karma start --chrome",
-		"lint": "eslint src",
+		"lint": "tslint -c tslint.json 'src/**/*.ts'",
 		"jsdoc": "rm -rf ./doc && jsdoc -c jsdoc.json",
 		"release": "node config/release.js",
 		"coverage": "karma start --coverage",
@@ -43,29 +43,20 @@
 	},
 	"homepage": "https://naver.github.io/egjs-component",
 	"devDependencies": {
-		"@babel/core": "^7.1.2",
-		"@babel/plugin-proposal-class-properties": "^7.1.0",
-		"@babel/preset-env": "^7.1.0",
+		"@egjs/build-helper": "^0.1.2",
 		"@egjs/common-demo": "github:naver/egjs#common-demo",
-		"babel-eslint": "^10.0.1",
-		"babel-loader": "^8.0.4",
-		"babel-plugin-no-side-effect-class-properties": "0.0.4",
-		"babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
-		"babel-plugin-transform-es3-property-literals": "^6.22.0",
+		"@types/chai": "^4.2.12",
+		"@types/karma-chai": "^0.1.2",
+		"@types/mocha": "^8.0.3",
 		"chai": "^4.1.1",
 		"chalk": "^2.1.0",
-		"clean-webpack-plugin": "^0.1.16",
 		"coveralls": "^2.13.1",
 		"cpx": "^1.5.0",
 		"egjs-jsdoc-template": "1.3.3",
-		"eslint": "^4.7.2",
-		"eslint-config-naver": "^1.0.1",
-		"eslint-loader": "^1.9.0",
 		"fs-extra": "^4.0.1",
 		"gh-pages": "^1.0.0",
 		"husky": "^0.14.3",
 		"inject-loader": "^3.0.1",
-		"istanbul-instrumenter-loader": "^3.0.0",
 		"jsdoc": "^3.5.4",
 		"karma": "^1.7.0",
 		"karma-chai": "^0.1.0",
@@ -74,24 +65,18 @@
 		"karma-mocha": "^1.3.0",
 		"karma-mocha-reporter": "^2.2.3",
 		"karma-sinon": "^1.0.5",
-		"karma-sourcemap-loader": "^0.3.7",
-		"karma-webpack": "^2.0.4",
 		"lite-fixture": "^1.0.2",
 		"marked": "^0.3.9",
 		"mocha": "^3.5.0",
+		"print-coveralls": "^1.2.2",
+		"print-sizes": "^0.1.0",
 		"rollup": "^0.66.4",
-		"rollup-plugin-babel": "^4.0.3",
-		"rollup-plugin-replace": "^2.1.0",
-		"rollup-plugin-uglify": "^6.0.0",
-		"sinon": "^3.2.1",
-		"string-replace-webpack-plugin": "0.1.3",
-		"sync-exec": "^0.6.2",
+		"sinon": "^3.3.0",
 		"tslint": "^6.1.3",
-		"uglifyjs-webpack-plugin": "^0.4.6",
-		"webpack": "^3.5.5",
-		"webpack-dev-server": "^2.7.1",
-		"webpack-merge": "^4.1.0",
-		"write-file-webpack-plugin": "^4.1.0",
-		"xml2js": "^0.4.19"
+		"tslint-consistent-codestyle": "^1.16.0",
+		"tslint-eslint-rules": "^5.4.0",
+		"tslint-lines-between-class-members": "^1.3.6",
+		"typescript": "^4.0.3",
+		"karma-typescript": "^5.2.0"
 	}
 }
