+++ tsconfig.test.json
@@ -1,0 +1,21 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "module": "commonjs",
+    "noImplicitAny": false,
+    "strictNullChecks": false,
+    "noUnusedLocals": false,
+    "experimentalDecorators": true,
+    "types": [
+      "karma-chai",
+      "mocha",
+    ]
+  },
+  "include": [
+    "./src/**/*.ts",
+    "./test/**/*.ts"
+  ],
+  "exclude": [
+    "./node_modules/**/*.ts"
+  ]
+}
