+++ tslint.json
@@ -20,6 +20,8 @@
         "severity": "warning"
       },
       "curly": false,
+      "forin": false,
+      "no-unused-expression": false,
       "adjacent-overload-signatures": false,
       "no-bitwise": false,
       "max-line-length": false,
